Version 3.0.7
commit 17946e9e70af136123a3334c7634a2465c46dee6
Author: Nils Hörrmann <post@nilshoerrmann.de>
Date:   Sat Apr 6 12:41:12 2013 +0200

    Version 3.0.7

commit 7683e76354640343d866b10fdda67cda0685b416
Author: Brendan Abbott <brendan@bloodbone.ws>
Date:   Sat Apr 6 11:37:27 2013 +1000

    Check to ensure the Stage table exists before doing update/uninstall. Fixes #126

commit d05d6f621932bb060337120a84fcbfe673b60fac
Merge: a02d4fa cde9596
Author: Brendan Abbott <brendan@bloodbone.ws>
Date:   Sat Apr 6 11:05:19 2013 +1000

    Merge branch 'development' of git://github.com/hananils/datetime into development

commit cde959604724a0a4c599e1121f60c55bc287ae22
Merge: 6e4821f 2af7274
Author: Brendan Abbott <brendan@bloodbone.ws>
Date:   Fri Apr 5 18:04:26 2013 -0700

    Merge pull request #136 from kkriehl/development

    Requirement-check not working with symphony 2.3.2

commit a02d4fa47a4b9d04e9b51b3e7a7d1f87a7e431e1
Author: Brendan Abbott <brendan@bloodbone.ws>
Date:   Sat Apr 6 11:04:11 2013 +1000

    Fix the styling of invalid dates, fix error in the JS where visualise was a parameter but also a function name

commit b2056a1ad6b81772219f64215881332c1fe9c88f
Author: Brendan Abbott <brendan@bloodbone.ws>
Date:   Sat Apr 6 11:03:44 2013 +1000

    Use AjaxPage and set the content type to be application/json for the AJAX validation of dates

commit 0d711222981c313027d1291518a1909e51769214
Author: Brendan Abbott <brendan@bloodbone.ws>
Date:   Sat Apr 6 10:54:08 2013 +1000

    Correct the prepareExportValue method. Fix the output of ranges.

commit ea0054e7e8a53599c304af718113b00f04f16352
Author: Brendan Abbott <brendan@bloodbone.ws>
Date:   Sat Apr 6 10:20:46 2013 +1000

    Couple of method signature updates for the DateTime DS

commit 2af7274c184a0289d67bad8b7f198159ba1e0902
Author: Kristian Kriehl <kristian.kriehl@gmx.de>
Date:   Sat Apr 6 01:41:00 2013 +0200

    Fixed the check of the required-setting of the field when saving a section entry. Additionally changed the relational operator to type-comparing (===) in all conditionals to avoid other unforeseen effects and int-casted numeric field-settings at the same time if necessary.

commit 6e4821fab296521ae20add7d7771ef5bb98dcbc8
Author: Brendan Abbott <brendan@bloodbone.ws>
Date:   Tue Jan 22 21:45:18 2013 +1000

    Whitespace and signature update for the update() method

commit 4987faaebf26aed4cc84e02db276f5e00447f6ad
Merge: 1c2b965 2c395a2
Author: Brendan Abbott <brendan@bloodbone.ws>
Date:   Sun Jan 13 01:40:18 2013 -0800

    Merge pull request #128 from alexbirukov/development

    Add Russian language

commit 1c2b965142f8d7bcd300a722bdbe012965dbd78e
Author: Brendan Abbott <brendan@bloodbone.ws>
Date:   Sun Jan 13 19:22:49 2013 +1000

    Update commit to use FieldManager::saveSettings

commit 2c395a26a5577bbfeff2b636df500222df7c7a0e
Author: alexbirukov <lord_maloy@mail.ru>
Date:   Sun Jan 13 09:19:32 2013 +0600

    Add russian language

    Add russian language file

commit 20523be6e015013e66f2fca1ac34dca3d87836ed
Author: Brendan Abbott <brendan@bloodbone.ws>
Date:   Tue Jan 8 23:11:58 2013 +1000

    Update prepareImportValue function

commit 296d362d4270b7a947008f3a1a123fcf3f6371a5
Author: Brendan Abbott <brendan@bloodbone.ws>
Date:   Sat Aug 18 08:16:12 2012 +1000

    Correct the definition of the prepareImportValue function

+++ extension.driver.php
@@ -7,7 +7,7 @@
 	 * Date and Time Extension
 	 */
 	Class extension_datetime extends Extension {
-	
+
 		private $languages = array(
 			'english' => 'en, en_GB.UTF8, en_GB',
 			'finnish' => 'fi, fi_FI.UTF8, fi_FI',
@@ -42,25 +42,25 @@
 		 * Add site preferences
 		 */
 		public function __addPreferences($context) {
-					
+
 			// Get selected languages
 			$selection = Symphony::Configuration()->get('datetime');
 			if(empty($selection)) $selection = array();
-		
+
 			// Build default options
 			$options = array();
 			foreach($this->languages as $name => $codes) {
 				$options[$name] = array($name . '::' . $codes, (array_key_exists($name, $selection) ? true : false), __(ucfirst($name)));
 			}
-			
+
 			// Add custom options
 			foreach(array_diff_key($selection, $this->languages) as $name => $codes) {
 				$options[$name] = array($name . '::' . $codes, true, __(ucfirst($name)));
 			}
-			
+
 			// Sort options
-			ksort($options);			
-			
+			ksort($options);
+
 			// Add fieldset
 			$group = new XMLElement('fieldset', '<legend>' . __('Date and Time') . '</legend>', array('class' => 'settings'));
 			$select = Widget::Select('settings[datetime][]', $options, array('multiple' => 'multiple'));
@@ -75,13 +75,13 @@
 		 * Save preferences
 		 */
 		public function __savePreferences($context) {
-	
+
 			// Remove old selection
-			Symphony::Configuration()->remove('datetime');	
+			Symphony::Configuration()->remove('datetime');
 
 			// Get selection
 			$selection = $context['settings']['datetime'];
-			
+
 			// Prepare preferences
 			$context['settings']['datetime'] = array();
 			if(is_array($selection)) {
@@ -91,7 +91,7 @@
 				}
 			}
 		}
-				
+
 		/**
 		 * @see http://symphony-cms.com/learn/api/2.3/toolkit/extension/#install
 		 */
@@ -112,7 +112,7 @@
 				) ENGINE=MyISAM DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;"
 			);
 
-			// Add language strings to configuration			
+			// Add language strings to configuration
 			Symphony::Configuration()->set('english', $this->languages['english'], 'datetime');
 			Administration::instance()->saveConfig();
 
@@ -128,7 +128,7 @@
 		/**
 		 * @see http://symphony-cms.com/learn/api/2.3/toolkit/extension/#update
 		 */
-		public function update($previousVersion) {
+		public function update($previousVersion = null) {
 			$status = array();
 
 			// Get table columns
@@ -192,34 +192,37 @@
 
 			// Prior version 2.4
 			if(version_compare($previousVersion, '2.4', '<')) {
-				
+
 				// Move language codes to configuration
 				Symphony::Configuration()->set('english', $this->languages['english'], 'datetime');
 				Symphony::Configuration()->set('german', $this->languages['german'], 'datetime');
 				Administration::instance()->saveConfig();
 			}
-			
+
 			// Prior version 3.0
 			if(version_compare($previousVersion, '3.0', '<')) {
-			
+
 				// Add multiple setting
 				if(!in_array('multiple', $columns)) {
 					$status[] = Symphony::Database()->query(
 						"ALTER TABLE `tbl_fields_datetime` ADD `multiple` tinyint(1) DEFAULT '1'"
 					);
 				}
-				
+
 				// Transfer old Stage settings
 				$constructables = Symphony::Database()->fetchCol("field_id", "SELECT `field_id` FROM  `tbl_fields_stage` WHERE  `constructable` = 0");
 				if(!empty($constructables) && is_array($constructables)) {
 					Symphony::Database()->query("UPDATE `tbl_fields_datetime` SET `multiple` = 0 WHERE `field_id` IN (" . implode(',', $constructables) . ")");
 				}
-				
+
 				// Remove old Stage instances
-				Symphony::Database()->query("DELETE FROM `tbl_fields_stage` WHERE `context` = 'datetime'");
-				Symphony::Database()->query("DELETE FROM `tbl_fields_stage_sorting` WHERE `context` = 'datetime'");
+				$does_stage_exist = Symphony::Database()->fetchRow(0, "SHOW TABLES LIKE 'tbl_fields_stage'");
+				if(!empty($does_stage_exist)) {
+					Symphony::Database()->query("DELETE FROM `tbl_fields_stage` WHERE `context` = 'datetime'");
+					Symphony::Database()->query("DELETE FROM `tbl_fields_stage_sorting` WHERE `context` = 'datetime'");
+				}
 			}
-			
+
 			// Report status
 			if(in_array(false, $status, true)) {
 				return false;
@@ -235,15 +238,18 @@
 		public function uninstall() {
 
 			// Remove old Stage tables if they are empty
-			$count = Symphony::Database()->query("SELECT COUNT(*) FROM `tbl_fields_stage`");
-			if($count == 0) {
-				Symphony::Database()->query("DROP TABLE `tbl_fields_stage`");
-				Symphony::Database()->query("DROP TABLE `tbl_fields_stage_sorting`");
+			$does_stage_exist = Symphony::Database()->fetchRow(0, "SHOW TABLES LIKE 'tbl_fields_stage'");
+			if(!empty($does_stage_exist)) {
+				$count = Symphony::Database()->query("SELECT COUNT(*) FROM `tbl_fields_stage`");
+				if($count == 0) {
+					Symphony::Database()->query("DROP TABLE `tbl_fields_stage`");
+					Symphony::Database()->query("DROP TABLE `tbl_fields_stage_sorting`");
+				}
 			}
-			
+
 			// Drop date and time table
 			Symphony::Database()->query("DROP TABLE `tbl_fields_datetime`");
-			
+
 			// Remove language strings from configuration
 			Symphony::Configuration()->remove('datetime');
 		}
